+++ src/mechanisms.rs
@@ -12,104 +12,87 @@
 
 // TODO: rename to aes256-cbc-zero-iv
 #[cfg(feature = "aes256-cbc")]
-pub struct Aes256Cbc {}
+pub struct Aes256Cbc;
 #[cfg(feature = "aes256-cbc")]
 mod aes256cbc;
 
 #[cfg(feature = "chacha8-poly1305")]
-pub struct Chacha8Poly1305 {}
+pub struct Chacha8Poly1305;
 #[cfg(feature = "chacha8-poly1305")]
 mod chacha8poly1305;
 
 #[cfg(feature = "shared-secret")]
-pub struct SharedSecret {}
+pub struct SharedSecret;
 #[cfg(feature = "shared-secret")]
 mod shared_secret;
 
 #[cfg(feature = "ed255")]
-pub struct Ed255 {}
+pub struct Ed255;
 #[cfg(feature = "ed255")]
 mod ed255;
 
 #[cfg(feature = "hmac-blake2s")]
-pub struct HmacBlake2s {}
+pub struct HmacBlake2s;
 #[cfg(feature = "hmac-blake2s")]
 mod hmacblake2s;
 
 #[cfg(feature = "hmac-sha1")]
-pub struct HmacSha1 {}
+pub struct HmacSha1;
 #[cfg(feature = "hmac-sha1")]
 mod hmacsha1;
 
 #[cfg(feature = "hmac-sha256")]
-pub struct HmacSha256 {}
+pub struct HmacSha256;
 #[cfg(feature = "hmac-sha256")]
 mod hmacsha256;
 
 #[cfg(feature = "hmac-sha512")]
-pub struct HmacSha512 {}
+pub struct HmacSha512;
 #[cfg(feature = "hmac-sha512")]
 mod hmacsha512;
 
 #[cfg(feature = "p256")]
-pub struct P256 {}
+pub struct P256;
 #[cfg(feature = "p256")]
-pub struct P256Prehashed {}
+pub struct P256Prehashed;
 #[cfg(feature = "p256")]
 mod p256;
 
 #[cfg(feature = "p384")]
-pub struct P384 {}
+pub struct P384;
 #[cfg(feature = "p384")]
-pub struct P384Prehashed {}
+pub struct P384Prehashed;
 #[cfg(feature = "p384")]
 mod p384;
 
 #[cfg(feature = "p521")]
-pub struct P521 {}
+pub struct P521;
 #[cfg(feature = "p521")]
-pub struct P521Prehashed {}
+pub struct P521Prehashed;
 #[cfg(feature = "p521")]
 mod p521;
 
 #[cfg(feature = "sha256")]
-pub struct Sha256 {}
+pub struct Sha256;
 #[cfg(feature = "sha256")]
 mod sha256;
 
 #[cfg(feature = "tdes")]
-pub struct Tdes {}
+pub struct Tdes;
 #[cfg(feature = "tdes")]
 mod tdes;
 
 #[cfg(feature = "totp")]
-pub struct Totp {}
+pub struct Totp;
 #[cfg(feature = "totp")]
 mod totp;
 
 #[cfg(feature = "trng")]
-pub struct Trng {}
+pub struct Trng;
 #[cfg(feature = "trng")]
 mod trng;
 
 #[cfg(feature = "x255")]
-pub struct X255 {}
+pub struct X255;
 #[cfg(feature = "x255")]
 mod x255;
-
-// pub enum MechanismEnum {
-//     NotImplemented,
-//     Ed255(ed255::Ed255),
-//     P256(p256::P256),
-// }
-
-// use crate::types::Mechanism;
-// pub fn enum_to_type(mechanism: Mechanism) -> MechanismEnum {
-//     match mechanism {
-//         #[cfg(feature = "ed255")]
-//         Mechanism::Ed255 => MechanismEnum::Ed255(ed255::Ed255 {} ),
-//         #[cfg(feature = "p256")]
-//         Mechanism::P256 => MechanismEnum::P256(p256::P256 {} ),
-//         _ => MechanismEnum::NotImplemented,
-//     }
-// }