+++ src/mechanisms/hmacblake2s.rs
@@ -1,13 +1,14 @@
 use crate::api::{reply, request};
 use crate::error::Error;
 use crate::key;
-use crate::service::{DeriveKey, Sign};
+use crate::service::MechanismImpl;
 use crate::store::keystore::Keystore;
 use crate::types::Signature;
 
-impl DeriveKey for super::HmacBlake2s {
+impl MechanismImpl for super::HmacBlake2s {
     #[inline(never)]
     fn derive_key(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::DeriveKey,
     ) -> Result<reply::DeriveKey, Error> {
@@ -37,11 +38,13 @@ impl DeriveKey for super::HmacBlake2s {
 
         Ok(reply::DeriveKey { key })
     }
-}
 
-impl Sign for super::HmacBlake2s {
     #[inline(never)]
-    fn sign(keystore: &mut impl Keystore, request: &request::Sign) -> Result<reply::Sign, Error> {
+    fn sign(
+        &self,
+        keystore: &mut impl Keystore,
+        request: &request::Sign,
+    ) -> Result<reply::Sign, Error> {
         use blake2::Blake2s256;
         use hmac::{Mac, SimpleHmac};
         type HmacBlake2s = SimpleHmac<Blake2s256>;