+++ src/mechanisms/p521.rs
@@ -12,10 +12,7 @@ use super::{P521Prehashed, P521};
 use crate::{
     api::{reply, request},
     key,
-    service::{
-        Agree, DeriveKey, DeserializeKey, Exists, GenerateKey, SerializeKey, Sign, UnsafeInjectKey,
-        Verify,
-    },
+    service::MechanismImpl,
     store::keystore::Keystore,
     types::{KeyId, KeySerialization, SerializedKey, Signature, SignatureSerialization},
     Error,
@@ -52,8 +49,9 @@ fn to_sec1_bytes(public_key: &p521::PublicKey) -> heapless::Vec<u8, { SCALAR_SIZ
     encoded_point.as_bytes().try_into().unwrap()
 }
 
-impl Agree for P521 {
+impl MechanismImpl for P521 {
     fn agree(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::Agree,
     ) -> Result<reply::Agree, Error> {
@@ -87,10 +85,10 @@ impl Agree for P521 {
             shared_secret: key_id,
         })
     }
-}
-impl DeriveKey for P521 {
+
     #[inline(never)]
     fn derive_key(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::DeriveKey,
     ) -> Result<reply::DeriveKey, Error> {
@@ -108,10 +106,10 @@ impl DeriveKey for P521 {
 
         Ok(reply::DeriveKey { key: public_id })
     }
-}
-impl DeserializeKey for P521 {
+
     #[inline(never)]
     fn deserialize_key(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::DeserializeKey,
     ) -> Result<reply::DeserializeKey, Error> {
@@ -146,10 +144,10 @@ impl DeserializeKey for P521 {
 
         Ok(reply::DeserializeKey { key: public_id })
     }
-}
-impl SerializeKey for P521 {
+
     #[inline(never)]
     fn serialize_key(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::SerializeKey,
     ) -> Result<reply::SerializeKey, Error> {
@@ -181,10 +179,10 @@ impl SerializeKey for P521 {
 
         Ok(reply::SerializeKey { serialized_key })
     }
-}
-impl Exists for P521 {
+
     #[inline(never)]
     fn exists(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::Exists,
     ) -> Result<reply::Exists, Error> {
@@ -192,11 +190,13 @@ impl Exists for P521 {
         let exists = keystore.exists_key(key::Secrecy::Secret, Some(key::Kind::P521), &key_id);
         Ok(reply::Exists { exists })
     }
-}
 
-impl Sign for P521 {
     #[inline(never)]
-    fn sign(keystore: &mut impl Keystore, request: &request::Sign) -> Result<reply::Sign, Error> {
+    fn sign(
+        &self,
+        keystore: &mut impl Keystore,
+        request: &request::Sign,
+    ) -> Result<reply::Sign, Error> {
         let key_id = request.key;
 
         let secret_key = load_secret_key(keystore, &key_id)?;
@@ -222,39 +222,9 @@ impl Sign for P521 {
             signature: serialized_signature,
         })
     }
-}
-impl Sign for P521Prehashed {
-    #[inline(never)]
-    fn sign(keystore: &mut impl Keystore, request: &request::Sign) -> Result<reply::Sign, Error> {
-        let key_id = request.key;
 
-        let secret_key = load_secret_key(keystore, &key_id)?;
-        let signing_key = SigningKey::from(ecdsa::SigningKey::from(secret_key));
-        let signature: p521::ecdsa::Signature = signing_key
-            .sign_prehash_with_rng(keystore.rng(), &request.message)
-            .map_err(|_| Error::InvalidSerializedRequest)?;
-
-        // debug_now!("making signature");
-        let serialized_signature = match request.format {
-            SignatureSerialization::Asn1Der => {
-                let der = signature.to_der();
-                Signature::from_slice(der.as_bytes()).unwrap()
-            }
-            SignatureSerialization::Raw => Signature::from_slice(&signature.to_bytes()).unwrap(),
-            _ => {
-                return Err(Error::InvalidSerializationFormat);
-            }
-        };
-
-        // return signature
-        Ok(reply::Sign {
-            signature: serialized_signature,
-        })
-    }
-}
-
-impl UnsafeInjectKey for P521 {
     fn unsafe_inject_key(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::UnsafeInjectKey,
     ) -> Result<reply::UnsafeInjectKey, Error> {
@@ -279,11 +249,10 @@ impl UnsafeInjectKey for P521 {
             )
             .map(|key| reply::UnsafeInjectKey { key })
     }
-}
 
-impl Verify for P521 {
     #[inline(never)]
     fn verify(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::Verify,
     ) -> Result<reply::Verify, Error> {
@@ -306,10 +275,9 @@ impl Verify for P521 {
         let valid = verifying_key.verify(&request.message, &signature).is_ok();
         Ok(reply::Verify { valid })
     }
-}
 
-impl GenerateKey for P521 {
     fn generate_key(
+        &self,
         keystore: &mut impl Keystore,
         request: &request::GenerateKey,
     ) -> Result<reply::GenerateKey, Error> {
@@ -324,5 +292,39 @@ impl GenerateKey for P521 {
 
         // return handle
         Ok(reply::GenerateKey { key: key_id })
+    }
+}
+
+impl MechanismImpl for P521Prehashed {
+    #[inline(never)]
+    fn sign(
+        &self,
+        keystore: &mut impl Keystore,
+        request: &request::Sign,
+    ) -> Result<reply::Sign, Error> {
+        let key_id = request.key;
+
+        let secret_key = load_secret_key(keystore, &key_id)?;
+        let signing_key = SigningKey::from(ecdsa::SigningKey::from(secret_key));
+        let signature: p521::ecdsa::Signature = signing_key
+            .sign_prehash_with_rng(keystore.rng(), &request.message)
+            .map_err(|_| Error::InvalidSerializedRequest)?;
+
+        // debug_now!("making signature");
+        let serialized_signature = match request.format {
+            SignatureSerialization::Asn1Der => {
+                let der = signature.to_der();
+                Signature::from_slice(der.as_bytes()).unwrap()
+            }
+            SignatureSerialization::Raw => Signature::from_slice(&signature.to_bytes()).unwrap(),
+            _ => {
+                return Err(Error::InvalidSerializationFormat);
+            }
+        };
+
+        // return signature
+        Ok(reply::Sign {
+            signature: serialized_signature,
+        })
     }
 }